requirements.txt
setuptools==80.9.0
jmespath==0.10.0
addict==2.4.0
pycryptodome==3.23.0
packaging==24.2
sentry-sdk==2.29.1
tqdm==4.65.2
lmdb==1.6.2
lightning==2.5.1.post0
opendatalab==0.0.10
urllib3==1.26.20
pytorch-lightning==2.5.1.post0
cryptography==45.0.3
aliyun-python-sdk-core==2.16.0
requests==2.28.2
mmcv-full==1.7.2
ordered-set==4.1.0
setproctitle==1.3.6
rich==13.4.2
openmim==0.3.9
oss2==2.17.0
yapf==0.43.0
docker-pycreds==0.4.0
GitPython==3.1.44
smmap==5.0.2
colorama==0.4.6
wandb==0.19.11
openxlab==0.1.2
filelock==3.14.0
model-index==0.1.11
gitdb==4.0.12
torchmetrics==1.7.2
crcmod==1.7
aliyun-python-sdk-kms==2.16.5
zipp==3.21.0
flash_attn==2.4.2.dev3
einops==0.8.0
importlib_metadata==8.5.0
transformer_engine==1.14.0+87fbe81
torch_tensorrt==2.6.0a0
apex==0.1
PuLP==2.9.0
polygraphy==0.49.16
torchprofile==0.0.4
safetensors==0.5.2
nvidia-modelopt==0.21.0
nvidia-pyindex==1.0.9
tensorrt==10.8.0.40
kvikio==24.10.0
dask-cuda==24.10.0
cuml==24.10.0
pylibcugraph==24.10.0
libcudf==24.10.0
zarr==2.18.3
distributed-ucxx==0.40.0
zict==3.0.0
aiohttp==3.11.10
aiohappyeyeballs==2.4.4
propcache==0.2.1
Pygments==2.18.0
cuda-python==12.6.2.post1
libucxx==0.40.0
partd==1.4.2
fastrlock==0.8.2
ply==3.11
asciitree==0.3.3
rapids-dask-dependency==24.10.0a0
frozenlist==1.5.0
thriftpy2==0.4.20
raft-dask==24.10.0
cupy-cuda12x==13.2.0
dask-cudf==24.10.0
cugraph-pyg==24.10.0
pylibwholegraph==24.10.0
pandas==2.2.2
tzdata==2024.2
toolz==1.0.0
psutil==6.1.0
locket==1.0.0
cudf==24.10.0
ucxx==0.40.0
cuvs==24.10.0
click==8.1.7
cugraph==24.10.0
aiosignal==1.3.2
fsspec==2024.10.0
pynvml==11.4.1
cachetools==5.5.0
multidict==6.1.0
grpcio==1.62.1
dask-expr==1.1.14
llvmlite==0.42.0
ucx-py==0.40.0
libkvikio==24.10.0
tblib==3.0.0
treelite==4.3.0
yarl==1.18.3
cugraph-equivariant==24.10.0
scipy==1.14.1
pylibcugraphops==24.10.0
numcodecs==0.13.1
dask==2024.9.0
cugraph-service-client==24.10.0
cugraph-service-server==24.10.0
pylibcudf==24.10.0
nx-cugraph==24.10.0
rmm==24.10.0
cloudpickle==3.1.0
six==1.17.0
fasteners==0.19
cudf-polars==24.10.0
distributed==2024.9.0
polars==1.8.2
librmm==24.10.0
pyarrow==17.0.0
nvtx==0.2.5
Jinja2==3.1.4
cugraph-dgl==24.10.0
xgboost==2.1.1
pytz==2023.4
attrs==24.2.0
pylibraft==24.10.0
numba==0.59.1
pynvjitlink==0.3.0
tabulate==0.9.0
nvidia-cudnn-frontend==1.9.0
onnx==1.17.0
pybind11_global==2.13.6
fonttools==4.55.3
kiwisolver==1.4.8
contourpy==1.3.1
pycocotools==2.0+nv0.8.1
matplotlib==3.10.0
cycler==0.12.1
pyparsing==3.2.1
nvidia-dali-cuda120==1.45.0
gast==0.6.0
dm-tree==0.1.8
nvidia-nvimgcodec-cu12==0.3.0.5
pytorch-triton==3.1.0+cf34004b8.internal
ninja==1.11.1.3
types-dataclasses==0.6.6
optree==0.14.0
lintrunner==0.12.7
expecttest==0.3.0
sympy==1.13.1
torchvision==0.20.0a0
pillow==11.1.0
torch==2.6.0a0+ecf3bae40a.nv25.1
networkx==3.4.2
mpmath==1.3.0
opencv==4.10.0
mdit-py-plugins==0.4.2
black==24.10.0
jupytext==1.16.6
isort==5.13.2
jupyterlab_tensorboard_pro==4.0.0
pathspec==0.12.1
mypy-extensions==1.0.0
pexpect==4.9.0
traitlets==5.14.3
httpx==0.28.1
jedi==0.19.2
soupsieve==2.6
bleach==6.2.0
jupyter-events==0.11.0
async-lru==2.0.4
pyzmq==26.2.0
types-python-dateutil==2.9.0.20241206
python-json-logger==3.2.1
argon2-cffi-bindings==21.2.0
stack-data==0.6.3
arrow==1.3.0
jsonschema-specifications==2024.10.1
ipython==8.31.0
httpcore==1.0.7
pandocfilters==1.5.1
pure_eval==0.2.3
rfc3339-validator==0.1.4
comm==0.2.2
executing==2.1.0
referencing==0.36.1
rfc3986-validator==0.1.1
jsonpointer==3.0.0
fqdn==1.5.1
nest-asyncio==1.6.0
Markdown==3.7
Werkzeug==3.1.3
argon2-cffi==23.1.0
fastjsonschema==2.21.1
uri-template==1.3.0
ipykernel==6.29.5
isoduration==20.11.0
webencodings==0.5.1
jupyterlab_code_formatter==3.0.2
jupyterlab==4.3.4
rpds-py==0.22.3
tinycss2==1.4.0
jupyter-lsp==2.2.5
sniffio==1.3.1
babel==2.16.0
python-hostlist==2.2.1
nbclient==0.10.2
Send2Trash==1.8.3
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
nbformat==5.10.4
debugpy==1.8.12
wcwidth==0.2.13
parso==0.8.4
defusedxml==0.7.1
asttokens==3.0.0
jupyter_server==2.15.0
tensorboard==2.16.2
json5==0.10.0
mistune==3.1.0
notebook_shim==0.2.4
jupyter_core==5.7.2
tensorboard-data-server==0.7.2
jupyter_server_terminals==0.5.3
overrides==7.7.0
terminado==0.18.1
notebook==7.3.2
jupyterlab_pygments==0.3.0
prometheus_client==0.21.1
jupyter_client==8.6.3
anyio==4.8.0
h11==0.14.0
nbconvert==7.16.5
tornado==6.4.2
matplotlib-inline==0.1.7
jupyterlab_server==2.27.3
webcolors==24.11.1
absl-py==2.1.0
ptyprocess==0.7.0
jsonschema==4.23.0
websocket-client==1.8.0
prompt_toolkit==3.0.48
xdoctest==1.0.2
pytest-flakefinder==1.1.0
pycparser==2.22
wheel==0.45.1
pooch==1.8.2
cloudpathlib==0.20.0
confection==0.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
blis==0.7.11
wasabi==1.1.3
execnet==2.1.1
annotated-types==0.7.0
markdown-it-py==3.0.0
pytest-xdist==3.6.1
astunparse==1.6.3
mkl-include==2021.1.1
PyYAML==6.0.2
soundfile==0.13.0
threadpoolctl==3.5.0
mkl-devel==2021.1.1
pluggy==1.5.0
thinc==8.2.5
pydantic==2.10.5
joblib==1.4.2
smart-open==7.1.0
weasel==0.4.1
cymem==2.0.11
hypothesis==5.35.1
spacy-legacy==3.0.12
pybind11==2.13.6
typing_extensions==4.12.2
pytest==8.1.1
pytest-rerunfailures==15.0
scikit-learn==1.6.1
protobuf==4.24.4
wrapt==1.17.2
decorator==5.1.1
charset-normalizer==3.4.1
numpy==1.26.4
mdurl==0.1.2
iniconfig==2.0.0
regex==2024.11.6
librosa==0.10.1
certifi==2024.12.14
shellingham==1.5.4
preshed==3.0.9
typer==0.15.1
catalogue==2.0.10
sortedcontainers==2.4.0
murmurhash==1.0.12
mock==5.1.0
soxr==0.5.0.post1
tbb==2021.13.1
audioread==3.0.1
srsly==2.5.1
idna==3.10
msgpack==1.1.0
mkl==2021.1.1
pytest-shard==0.1.2
platformdirs==4.3.6
pydantic_core==2.27.2
language_data==1.3.0
langcodes==3.5.0
cffi==1.17.1
intel-openmp==2021.4.0
marisa-trie==1.2.1
spacy==3.7.5
Cython==3.0.11
spacy-loggers==1.0.5
cmake==3.31.4
pip==24.3.1
nvfuser==0.2.23a0+6627725
lightning-thunder==0.2.0.dev0
dill==0.3.9
opt_einsum==3.4.0
lightning-utilities==0.12.0.dev0
looseversion==1.3.0
wheel==0.45.1
packaging==24.2
typeguard==4.3.0
platformdirs==4.2.2
more-itertools==10.3.0
tomli==2.0.1
inflect==7.3.1
jaraco.functools==4.0.1
typing_extensions==4.12.2
zipp==3.19.2
importlib_metadata==8.0.0
backports.tarfile==1.2.0
autocommand==2.2.2
jaraco.collections==5.1.0
jaraco.text==3.12.1
jaraco.context==5.3.0
